--- src/features/share-schedule/index.ts
@@ -1,7 +1,0 @@
-// API
-export * from './api/share-schedule.API'
-
-// UI
-export { default as ShareSchedule } from './share-calendar/share-calendar'
-export { default as ShareDetailedSchedule } from './detailed-share-schedule/ui/share-detailed-schedule'
-export { default as ShareDetailedScheduleList } from './detailed-share-schedule/ui/share-detailed-schedule-list'
