+++ src/features/share-schedule/ui/share-detailed-schedule.tsx
