+++ src/pages/detailed-schedule/share-detailed-schedule/share-detailed-schedule-list-page.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from 'react'
 
-import { ShareDetailedScheduleList } from '@/features/share-schedule'
+import { ShareDetailedScheduleList } from '@/features/share-schedule/ui'
 
 import DetailedScheduleListSkeleton from '../detailed-schedule-list-skeleton'
